Refactor the events.acquisition and components package, fixed waning on MqttCallback
