+++ src/main/java/dtb/lego/truck/rest/component/entity/Component.java
@@ -1,0 +1,11 @@
+package dtb.lego.truck.rest.component.entity;
+
+/**
+ * Used to define the types of Component available.
+ */
+public enum Component {
+    ACCELEROMETER,
+    GYROSCOPE,
+    MOTOR,
+    PROXIMITY
+}
