+++ src/main/java/dtb/lego/truck/rest/component/entity/Transformation.java
@@ -1,0 +1,14 @@
+package dtb.lego.truck.rest.component.entity;
+
+/**
+ * Contains the implemented transformations available to downsample the data
+ */
+public enum Transformation {
+    LAST,
+    MEAN,
+    MAX,
+    MIN,
+    MEDIAN
+
+
+}
