--- src/main/java/dtb/lego/truck/rest/component/events/entity/Component.java
@@ -1,11 +1,0 @@
-package dtb.lego.truck.rest.component.events.entity;
-
-/**
- * Used to define the types of Component available.
- */
-public enum Component {
-    ACCELEROMETER,
-    GYROSCOPE,
-    MOTOR,
-    PROXIMITY
-}
