--- src/main/java/dtb/lego/truck/rest/component/events/entity/Transformation.java
@@ -1,14 +1,0 @@
-package dtb.lego.truck.rest.component.events.entity;
-
-/**
- * Contains the implemented transformations available to downsample the data
- */
-public enum Transformation {
-    LAST,
-    MEAN,
-    MAX,
-    MIN,
-    MEDIAN
-
-
-}
