--- src/main/java/dtb/lego/truck/rest/component/events/entity/repositories/ProximitySensorEventRepository.java
@@ -1,13 +1,0 @@
-package dtb.lego.truck.rest.component.events.entity.repositories;
-
-import dtb.lego.truck.rest.component.events.entity.events.ProximitySensorEvent;
-
-import javax.transaction.Transactional;
-
-/**
- * CrudRepository extension for Gyroscope Events.
- * Used in order to store the gyroscope events on the database
- */
-@Transactional
-public interface ProximitySensorEventRepository extends EventRepository<ProximitySensorEvent> {
-}
